# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Go语言开发的八字分析系统，提供完整的6步分析流程",
        "title": "八字分析系统 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/bazi/analyze": {
            "get": {
                "description": "输入八字四柱，返回完整6步分析结果。支持JSON和Markdown格式，通过Accept头或format参数控制",
                "produces": [
                    "application/json",
                    "text/markdown"
                ],
                "tags": [
                    "八字分析"
                ],
                "summary": "八字分析接口（GET方式）",
                "parameters": [
                    {
                        "type": "string",
                        "example": "壬戌 壬寅 庚午 丙戌",
                        "description": "八字四柱，格式：年柱 月柱 日柱 时柱",
                        "name": "bazi",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "json",
                            "markdown"
                        ],
                        "type": "string",
                        "default": "json",
                        "description": "返回格式：json或markdown",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BaZiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.BaZiResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "输入八字四柱，返回完整6步分析结果。支持JSON和Markdown格式，通过Accept头或format参数控制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/markdown"
                ],
                "tags": [
                    "八字分析"
                ],
                "summary": "八字分析接口（POST方式）",
                "parameters": [
                    {
                        "description": "八字请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BaZiRequest"
                        }
                    },
                    {
                        "enum": [
                            "json",
                            "markdown"
                        ],
                        "type": "string",
                        "default": "json",
                        "description": "返回格式：json或markdown",
                        "name": "format",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BaZiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.BaZiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BaZiRequest": {
            "type": "object",
            "required": [
                "bazi"
            ],
            "properties": {
                "bazi": {
                    "type": "string",
                    "example": "壬戌 壬寅 庚午 丙戌"
                }
            }
        },
        "api.BaZiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}